replace redis store with exact one
also, use a different default prefix
