+++ lib/ablab/store.rb
@@ -1,3 +1,2 @@
 require 'ablab/store/memory'
 require 'ablab/store/redis'
-require 'ablab/store/redis_set'
