+++ spec/ablab/store/redis_spec.rb
@@ -21,5 +21,17 @@ describe Ablab::Store::Redis do
 
   let(:store) { Ablab::Store::Redis.new(db: 2, key_prefix: 'ablabtest') }
   include_examples 'store'
+
+  describe 'spool_set!' do
+    it 'does not change the count' do
+      5.times do |i|
+        store.track_view!('foo', 'bar', i)
+      end
+      later = Time.at(Time.now.to_i + 60 * 60)
+      allow(Time).to receive(:now).and_return(later)
+      expect { store.send(:spool_set!, 'foo', 'bar', :sessions) }
+        .not_to change { store.sessions('foo', 'bar') }
+    end
+  end
 end
